pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-research-engineer"
version = "0.1.0"
description = "Paper-to-blueprint compiler with feasibility gate for metis-platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "agent-factors",
]

[project.optional-dependencies]
graph = ["networkx>=3.0"]
integration = ["prior-art-retrieval"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "prior-art-retrieval",
]

[tool.setuptools.packages.find]
include = ["research_engineer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
